Refactor RAG modules
