+++ src/trackrealties/rag/optimized_pipeline.py
@@ -13,7 +13,7 @@ from src.trackrealties.core.graph import graph_manager
 from src.trackrealties.core.database import db_pool
 from src.trackrealties.models.search import SearchResult
 from src.trackrealties.analytics.search import search_analytics, SearchAnalytics
-from smart_search_implementation import (
+from .smart_search import (
     SmartSearchRouter,
     FixedGraphSearch,
     SearchStrategy,
@@ -598,5 +598,5 @@ class HybridSearchEngine(EnhancedRAGPipeline):
     pass
 
 # Or update your existing imports:
-from .optimized_search import EnhancedRAGPipeline as HybridSearchEngine
+from .optimized_pipeline import EnhancedRAGPipeline as HybridSearchEngine
 """
